requirements.txt
fastapi
uvicorn
httpx
watchfilespytest
httpx
